Using RulesIniSectionBuilder to split out code
