+++ TIBERIANDAWN/TiberianDawn.vcxproj
@@ -465,7 +465,7 @@
     <ClInclude Include="RULES.H" />
     <ClInclude Include="RulesIniRuleKey.h" />
     <ClInclude Include="RulesIniSection.h" />
-    <ClInclude Include="RulesIniTemplates.h" />
+    <ClInclude Include="RulesIniSectionBuilder.h" />
     <ClInclude Include="RulesIniTypes.h" />
     <ClInclude Include="RULES_CACHE_KEY.H" />
     <ClInclude Include="RULES_INI.H" />
